chore: address first cut reviews
